requirements.txt
Flask==2.3.3
nltk==3.8.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
python-dotenv==1.0.0
transformers==4.31.0
torch==2.0.1
pdfplumber==0.10.2
gunicorn==21.2.0